# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Prediction League 26/27",
    "description": "Premier League predictions, FPL and a watch log for four friends. Every response is served from cache; the poller is the only process that talks to an upstream.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/fixtures": {
      "get": {
        "tags": [
          "football"
        ],
        "summary": "Fixtures",
        "operationId": "fixtures_api_fixtures_get",
        "parameters": [
          {
            "name": "from",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "From"
            }
          },
          {
            "name": "to",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "To"
            }
          },
          {
            "name": "gameweek",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Gameweek"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FixtureListOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/fixtures/{fixture_id}": {
      "get": {
        "tags": [
          "football"
        ],
        "summary": "Fixture",
        "operationId": "fixture_api_fixtures__fixture_id__get",
        "parameters": [
          {
            "name": "fixture_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Fixture Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FixtureOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/predictions": {
      "get": {
        "tags": [
          "predictions"
        ],
        "summary": "List Predictions",
        "operationId": "list_predictions_api_predictions_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PredictionsOut"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "predictions"
        ],
        "summary": "Put Prediction",
        "description": "Owner only, and only before the lock.",
        "operationId": "put_prediction_api_predictions_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PredictionIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PredictionOut"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          }
        }
      }
    },
    "/api/predictions/preview": {
      "post": {
        "tags": [
          "predictions"
        ],
        "summary": "Preview",
        "description": "Score a draft table against a finished season.\n\nThe picker calls this rather than reimplementing the rules in TypeScript --\ntwo implementations of the scoring rules is a bug waiting to happen.",
        "operationId": "preview_api_predictions_preview_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PreviewIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PreviewOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ClubOut": {
        "properties": {
          "short_name": {
            "type": "string",
            "title": "Short Name"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "full_name": {
            "type": "string",
            "title": "Full Name"
          },
          "primary": {
            "type": "string",
            "title": "Primary"
          },
          "on_primary": {
            "type": "string",
            "title": "On Primary"
          },
          "fpl_id": {
            "type": "integer",
            "title": "Fpl Id"
          }
        },
        "type": "object",
        "required": [
          "short_name",
          "name",
          "full_name",
          "primary",
          "on_primary",
          "fpl_id"
        ],
        "title": "ClubOut",
        "description": "A club, with everything a crest needs and nothing it does not."
      },
...
      "ErrorOut": {
        "properties": {
          "detail": {
            "type": "string",
            "title": "Detail"
          }
        },
        "type": "object",
        "required": [
          "detail"
        ],
        "title": "ErrorOut"
      },
      "FixtureListOut": {
        "properties": {
          "fixtures": {
            "items": {
              "$ref": "#/components/schemas/FixtureOut"
            },
            "type": "array",
            "title": "Fixtures"
          },
          "freshness": {
            "$ref": "#/components/schemas/Freshness"
          },
          "empty_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Empty Message"
          }
        },
        "type": "object",
        "required": [
          "fixtures",
          "freshness",
          "empty_message"
        ],
        "title": "FixtureListOut"
      },
      "FixtureOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "gameweek": {
            "type": "integer",
            "title": "Gameweek"
          },
          "kickoff": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Kickoff"
          },
          "home": {
            "$ref": "#/components/schemas/ClubOut"
          },
          "away": {
            "$ref": "#/components/schemas/ClubOut"
          },
          "home_score": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Home Score"
          },
          "away_score": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Away Score"
          },
          "started": {
            "type": "boolean",
            "title": "Started",
            "default": false
          },
          "finished": {
            "type": "boolean",
            "title": "Finished",
            "default": false
          },
          "postponed": {
            "type": "boolean",
            "title": "Postponed",
            "default": false
          },
          "minutes": {
            "type": "integer",
            "title": "Minutes",
            "default": 0
          },
          "local_times": {
            "items": {
              "$ref": "#/components/schemas/LocalTimeOut"
            },
            "type": "array",
            "title": "Local Times"
          },
          "odds": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/OddsPrice"
              },
              {
                "type": "null"
              }
            ]
          },
          "derby": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Derby"
          },
          "watched_by": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Watched By"
          },
          "watch_open": {
            "type": "boolean",
            "title": "Watch Open",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "id",
          "gameweek",
          "kickoff",
          "home",
          "away",
          "home_score",
          "away_score",
          "started",
          "finished",
          "postponed",
          "minutes",
          "local_times",
          "odds",
          "derby",
          "watched_by",
          "watch_open"
        ],
        "title": "FixtureOut"
      },
      "Freshness": {
        "properties": {
          "source": {
            "type": "string",
            "title": "Source"
          },
          "age_seconds": {
            "type": "number",
            "title": "Age Seconds"
          },
          "stale": {
            "type": "boolean",
            "title": "Stale"
          },
          "available": {
            "type": "boolean",
            "title": "Available",
            "default": true
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          }
        },
        "type": "object",
        "required": [
          "source",
          "age_seconds",
          "stale",
          "available",
          "reason"
        ],
        "title": "Freshness",
        "description": "How old a cached payload is, so a panel can say so rather than lie."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LocalTimeOut": {
        "properties": {
          "place": {
            "type": "string",
            "title": "Place"
          },
          "person": {
            "type": "string",
            "title": "Person"
          },
          "city": {
            "type": "string",
            "title": "City"
          },
          "timezone": {
            "type": "string",
            "title": "Timezone"
          },
          "iso": {
            "type": "string",
            "format": "date-time",
            "title": "Iso"
          },
          "time": {
            "type": "string",
            "title": "Time"
          },
          "weekday": {
            "type": "string",
            "title": "Weekday"
          },
          "day": {
            "type": "string",
            "title": "Day"
          },
          "offset": {
            "type": "string",
            "title": "Offset"
          },
          "abbreviation": {
            "type": "string",
            "title": "Abbreviation"
          },
          "is_night": {
            "type": "boolean",
            "title": "Is Night"
          },
          "day_shift": {
            "type": "integer",
            "title": "Day Shift"
          },
          "broadcaster": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Broadcaster"
          },
          "broadcaster_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Broadcaster Url"
          },
          "verified_on": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Verified On"
          }
        },
        "type": "object",
        "required": [
          "place",
          "person",
          "city",
          "timezone",
          "iso",
          "time",
          "weekday",
          "day",
          "offset",
          "abbreviation",
          "is_night",
          "day_shift",
          "broadcaster",
          "broadcaster_url",
          "verified_on"
        ],
        "title": "LocalTimeOut",
        "description": "One kickoff rendered for one city."
      },
...
      "PredictionIn": {
        "properties": {
          "table": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Table"
          },
          "awards": {
            "$ref": "#/components/schemas/AwardPicks-Input"
          },
          "champions_league": {
            "$ref": "#/components/schemas/ChampionsLeaguePicks-Input"
          }
        },
        "type": "object",
        "title": "PredictionIn"
      },
      "PredictionOut": {
        "properties": {
          "person": {
            "type": "string",
            "title": "Person"
          },
          "filed": {
            "type": "boolean",
            "title": "Filed"
          },
          "redacted": {
            "type": "boolean",
            "title": "Redacted",
            "default": false
          },
          "table": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Table"
          },
          "awards": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AwardPicks-Output"
              },
              {
                "type": "null"
              }
            ]
          },
          "champions_league": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ChampionsLeaguePicks-Output"
              },
              {
                "type": "null"
              }
            ]
          },
          "submitted_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Submitted At"
          },
          "locked": {
            "type": "boolean",
            "title": "Locked",
            "default": false
          },
          "status": {
            "type": "string",
            "enum": [
              "filed",
              "open",
              "did-not-file"
            ],
            "title": "Status",
            "default": "open"
          }
        },
        "type": "object",
        "required": [
          "person",
          "filed",
          "redacted",
          "table",
          "awards",
          "champions_league",
          "submitted_at",
          "locked",
          "status"
        ],
        "title": "PredictionOut"
      },
      "PredictionsOut": {
        "properties": {
          "predictions": {
            "items": {
              "$ref": "#/components/schemas/PredictionOut"
            },
            "type": "array",
            "title": "Predictions"
          },
          "locked": {
            "type": "boolean",
            "title": "Locked"
          },
          "lock_at": {
            "type": "string",
            "format": "date-time",
            "title": "Lock At"
          },
          "seconds_remaining": {
            "type": "number",
            "title": "Seconds Remaining"
          }
        },
        "type": "object",
        "required": [
          "predictions",
          "locked",
          "lock_at",
          "seconds_remaining"
        ],
        "title": "PredictionsOut"
      },
      "PreviewIn": {
        "properties": {
          "table": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Table"
          },
          "awards": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AwardPicks-Input"
              },
              {
                "type": "null"
              }
            ]
          },
          "against_season": {
            "type": "string",
            "title": "Against Season",
            "default": "2025-26"
          }
        },
        "type": "object",
        "required": [
          "table"
        ],
        "title": "PreviewIn"
      },
      "PreviewOut": {
        "properties": {
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "table_points": {
            "type": "integer",
            "title": "Table Points"
          },
          "award_points": {
            "type": "integer",
            "title": "Award Points"
          },
          "exact_hits": {
            "type": "integer",
            "title": "Exact Hits"
          },
          "near_hits": {
            "type": "integer",
            "title": "Near Hits"
          },
          "top_four_bonus": {
            "type": "integer",
            "title": "Top Four Bonus"
          },
          "champion_bonus": {
            "type": "integer",
            "title": "Champion Bonus"
          },
          "against_season": {
            "type": "string",
            "title": "Against Season"
          },
          "per_club": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Per Club"
          }
        },
        "type": "object",
        "required": [
          "total",
          "table_points",
          "award_points",
          "exact_hits",
          "near_hits",
          "top_four_bonus",
          "champion_bonus",
          "against_season",
          "per_club"
        ],
        "title": "PreviewOut"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}